requirements.txt
# === Core Agent Framework ===
google-adk>=0.2.0

# === LLM Routing ===
litellm>=1.44.0

# === Chat UI ===
chainlit>=1.3.0

# === Web Server (landing page) ===
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9

# === Vector Store (handles embeddings internally via ONNX) ===
chromadb>=0.5.0

# === Config & Validation ===
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1

# === CLI ===
click>=8.1.7

# === Logging ===
structlog>=24.2.0

# === Dev / Test ===
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-cov>=5.0.0
ruff>=0.4.0
